@@ -1,106 +1,109 @@
 {
-  "name": "react-base",
-  "version": "1.0.0",
-  "description": "",
+  "name": "react-starter-kit",
+  "version": "2.0.0",
+  "description": "react starter kit",
   "main": "index.js",
   "scripts": {
-    "start": "webpack-dev-server --config ./scripts/webpack.dev.config.js",
-    "build": "webpack --config ./scripts/webpack.prod.config.js",
-    "createDllDev": "webpack --env.mode=development --config ./scripts/webpack.dll.config.js",
-    "createDllPro": "webpack --env.mode=production --config ./scripts/webpack.dll.config.js",
+    "start": "webpack serve --color --config ./scripts/webpack.dev.config.js",
+    "build": "webpack --color --config ./scripts/webpack.prod.config.js",
+    "createDllDev": "webpack --color --env development --config ./scripts/webpack.dll.config.js",
+    "createDllPro": "webpack --color --env production --config ./scripts/webpack.dll.config.js",
     "createDll": "npm run createDllDev && npm run createDllPro",
     "test:js": "jest --config=jest-js.config.js",
     "test:ui": "jest --config=jest-ui.config.js --runInBand",
     "huskyInstall": "node ./node_modules/husky/lib/installer/bin.js install"
   },
-  "author": "",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/One-boy/react-starter-kit.git"
+  },
+  "keywords": [
+    "react",
+    "starter",
+    "kit"
+  ],
+  "author": "huyu",
   "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/One-boy/react-starter-kit/issues"
+  },
+  "homepage": "https://github.com/One-boy/react-starter-kit#readme",
   "devDependencies": {
-    "@babel/cli": "^7.8.4",
-    "@babel/core": "^7.9.0",
-    "@babel/plugin-proposal-class-properties": "^7.8.3",
-    "@babel/plugin-proposal-decorators": "^7.8.3",
+    "@babel/cli": "^7.13.14",
+    "@babel/core": "^7.13.15",
+    "@babel/plugin-proposal-class-properties": "^7.13.0",
+    "@babel/plugin-proposal-decorators": "^7.13.15",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/preset-env": "^7.9.5",
-    "@babel/preset-react": "^7.9.4",
-    "acorn": "^6.4.1",
-    "autoprefixer": "^9.7.6",
+    "@babel/plugin-transform-runtime": "^7.13.15",
+    "@babel/preset-env": "^7.13.15",
+    "@babel/preset-react": "^7.13.13",
+    "autoprefixer": "^10.2.5",
     "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.1.0",
-    "babel-plugin-import": "^1.13.0",
-    "clean-webpack-plugin": "^2.0.1",
-    "copy-webpack-plugin": "^5.1.1",
-    "css-loader": "^0.28.11",
-    "cssnano": "^4.1.10",
+    "babel-loader": "^8.2.2",
+    "babel-plugin-import": "^1.13.3",
+    "clean-webpack-plugin": "^4.0.0-alpha.0",
+    "copy-webpack-plugin": "^8.1.1",
+    "css-loader": "^5.2.1",
     "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
-    "eslint": "^4.19.1",
-    "eslint-config-standard": "^11.0.0",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-node": "^6.0.1",
-    "eslint-plugin-promise": "^3.8.0",
-    "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-standard": "^3.1.0",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "^1.1.11",
+    "enzyme-adapter-react-16": "^1.15.6",
+    "eslint": "^7.24.0",
+    "eslint-config-prettier": "^8.1.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-plugin-react": "^7.23.2",
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "file-loader": "^6.2.0",
     "happypack": "^5.0.1",
-    "html-webpack-plugin": "^3.2.0",
-    "husky": "^2.7.0",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^24.9.0",
-    "jest-environment-enzyme": "^7.1.2",
-    "jest-enzyme": "^7.1.2",
+    "html-webpack-plugin": "^5.3.1",
+    "husky": "^6.0.0",
+    "jest": "^26.6.3",
     "jest-puppeteer": "^4.4.0",
-    "less": "^3.11.1",
-    "less-loader": "^4.1.0",
-    "lint-staged": "^8.2.1",
-    "mini-css-extract-plugin": "^0.4.5",
+    "less": "^4.1.1",
+    "less-loader": "^8.1.0",
+    "lint-staged": "^10.5.4",
+    "mini-css-extract-plugin": "^1.4.1",
     "open-browser-webpack-plugin": "0.0.5",
-    "postcss": "^7.0.27",
-    "postcss-loader": "^3.0.0",
-    "puppeteer": "^3.0.2",
-    "style-loader": "^0.20.3",
-    "url-loader": "^1.1.2",
-    "webpack": "^4.43.0",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.10.3",
-    "webpack-merge": "^4.2.2",
-    "worker-loader": "^2.0.0"
+    "postcss": "^8.2.10",
+    "postcss-loader": "^5.2.0",
+    "prettier": "2.2.1",
+    "puppeteer": "^8.0.0",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.32.0",
+    "webpack-bundle-analyzer": "^4.4.0",
+    "webpack-cli": "^4.6.0",
+    "webpack-dev-server": "^3.11.2",
+    "webpack-merge": "^5.7.3",
+    "worker-loader": "^3.0.8"
   },
   "dependencies": {
-    "@babel/polyfill": "^7.8.7",
-    "@hot-loader/react-dom": "^17.0.0",
-    "antd": "^3.26.16",
-    "axios": "^0.18.1",
-    "blob-polyfill": "^2.0.20171115",
-    "formdata-polyfill": "^3.0.19",
+    "@ant-design/icons": "^4.6.2",
+    "@babel/runtime": "^7.13.10",
+    "@hot-loader/react-dom": "^17.0.1",
+    "antd": "^4.15.1",
+    "axios": "^0.21.1",
+    "core-js": "^3.10.1",
+    "moment": "^2.29.1",
     "prop-types": "^15.7.2",
-    "raf": "^3.4.1",
-    "rc-animate": "^2.11.1",
-    "rc-banner-anim": "^2.4.3",
-    "rc-queue-anim": "^1.8.4",
-    "rc-scroll-anim": "^2.7.4",
-    "rc-texty": "^0.2.0",
-    "rc-tween-one": "^2.6.8",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-hot-loader": "^4.13.0",
-    "react-router-dom": "^5.1.2",
-    "webpack-bundle-analyzer": "^3.7.0"
-  },
-  "lint-staged": {
-    "linters": {
-      "app/**/*.js": [
-        "./node_modules/.bin/eslint --fix",
-        "git add"
-      ]
-    }
+    "react-router-dom": "^5.2.0"
   },
   "husky": {
     "hooks": {
-      "pre-commit": "lint-staged",
-      "commit-msg": "echo $HUSKY_GIT_PARAMS"
+      "pre-commit": "lint-staged"
     }
+  },
+  "lint-staged": {
+    "src/*.{js,jsx,mjs,ts,tsx}": [
+      "node_modules/.bin/prettier --write",
+      "node_modules/.bin/eslint --fix",
+      "git add"
+    ],
+    "src/*.{css,scss,less,json,html,md,markdown}": [
+      "node_modules/.bin/prettier --write",
+      "git add"
+    ]
   }
 }
