@@ -6,33 +6,29 @@
 module.exports = {
   env: {
     browser: true,
-    es6: true
+    es6: true,
+    node: true,
   },
-  parser: "babel-eslint",
-  extends: ["eslint:recommended", "plugin:react/recommended"],
+  // 使用babel开发的解析器
+  parser: 'babel-eslint',
+  extends: ['eslint:recommended', 'plugin:react/recommended', 'prettier'],
   parserOptions: {
     ecmaFeatures: {
       experimentalObjectRestSpread: true,
       jsx: true,
-      legacyDecorators: true
+      legacyDecorators: true,
     },
-    sourceType: "module"
+    sourceType: 'module',
   },
-  plugins: ["react", "jsx-a11y"],
+  plugins: ['react', 'jsx-a11y', 'prettier'],
   rules: {
-    "no-console": 0,
-    "no-debugger": 0,
-    "react/display-name": 0, //防止在React组件定义中丢失displayName
-    indent: [
-      "error",
-      2,
-      {
-        SwitchCase: 1
-      }
-    ],
-    "linebreak-style": 0,
-    quotes: ["error", "single"],
-    semi: ["error", "never"],
-    "spaced-comment": "error"
-  }
-};
+    'no-console': 0,
+    'no-debugger': 0,
+    // 防止在React组件定义中丢失displayName
+    'react/display-name': 0,
+    // prettier 标记的地方抛出错误信息
+    'prettier/prettier': 'error',
+    'arrow-body-style': 'off',
+    'prefer-arrow-callback': 'off',
+  },
+}
