@@ -1,7 +1,7 @@
-
 import User from './User'
 import List from './List'
 
 export default {
-  User, List,
+  User,
+  List,
 }
