@@ -2,7 +2,7 @@
  * @Author: huyu
  * @Date: 2019-05-28 10:12:29
  * @Last Modified by: huyu
- * @Last Modified time: 2019-05-29 14:12:58
+ * @Last Modified time: 2021-04-13 19:01:57
  */
 
 // 错误边界组件
@@ -13,7 +13,6 @@ import React, { Component } from 'react'
 import { Row, Col, Button, Card } from 'antd'
 import Proptypes from 'prop-types'
 import { withRouter } from 'react-router-dom'
-
 
 @withRouter
 export default class ErrorBoundary extends Component {
@@ -28,14 +27,13 @@ export default class ErrorBoundary extends Component {
     children: Proptypes.object,
   }
 
-
   static getDerivedStateFromError(error) {
     return { hasError: true }
   }
 
   componentDidCatch(error, info) {
     // You can also log the error to an error reporting service
-    // 
+    //
     console.error(error)
   }
 
@@ -58,11 +56,20 @@ export default class ErrorBoundary extends Component {
     if (this.state.hasError) {
       return (
         <Card bordered={false}>
-          <Row >
-            <Col span={24}> <p>Something has error</p></Col>
-            <Col span={6}><Button onClick={this.onClickGoBack}>回到上一页</Button></Col>
-            <Col span={6}><Button onClick={this.onClickGoHome}>去首页</Button></Col>
-            <Col span={6}><Button onClick={this.onClickReload}>重新加载</Button></Col>
+          <Row>
+            <Col span={24}>
+              {' '}
+              <p>Something has error</p>
+            </Col>
+            <Col span={6}>
+              <Button onClick={this.onClickGoBack}>回到上一页</Button>
+            </Col>
+            <Col span={6}>
+              <Button onClick={this.onClickGoHome}>去首页</Button>
+            </Col>
+            <Col span={6}>
+              <Button onClick={this.onClickReload}>重新加载</Button>
+            </Col>
           </Row>
         </Card>
       )
