@@ -1,7 +1,4 @@
 import Provider from './provider'
 import Connect from './connect'
 
-export {
-  Provider,
-  Connect,
-}
+export { Provider, Connect }
