@@ -1,10 +1,5 @@
-
 // 登录页测试
-import {
-  TIMEOUT,
-  LOG_DIR,
-  URL_PREFIX,
-} from './base'
+import { TIMEOUT, LOG_DIR, URL_PREFIX } from './base'
 
 describe('Login', () => {
   beforeAll(async () => {
@@ -25,7 +20,11 @@ describe('Login', () => {
     await page.type('#password', '1')
     await page.click('button[type="submit"]')
     // 截图
-    await page.screenshot({ path: `${LOG_DIR}/login.png`, type: 'png', fullPage: true })
+    await page.screenshot({
+      path: `${LOG_DIR}/login.png`,
+      type: 'png',
+      fullPage: true,
+    })
     const text = await page.evaluate(() => document.body.innerHTML)
     expect(text).toContain('中后台管理系统') // 登录成功并跳转后，页面应该包含该文本
   })
