@@ -1,8 +1,8 @@
 /*
- * @Author: huyu 
- * @Date: 2019-05-30 09:52:32 
+ * @Author: huyu
+ * @Date: 2019-05-30 09:52:32
  * @Last Modified by: huyu
- * @Last Modified time: 2019-05-30 11:10:58
+ * @Last Modified time: 2021-04-13 16:03:47
  */
 
 // 使用worker-loader
@@ -10,5 +10,7 @@
 
 onmessage = (e) => {
   console.log('Worker:收到Main发过来的消息', e.data)
-  postMessage(`Worker已收到来自Main的消息: ${e.data} --${new Date().toLocaleString()}`)
+  postMessage(
+    `Worker已收到来自Main的消息: ${e.data} --${new Date().toLocaleString()}`
+  )
 }
